PROTON-1512: Fixes and doc clarifications for pn_delivery_aborted()
- added explict PN_ABORTED error code (PN_STATE_ERR has another meaning in this context)
- fixed initialization bug for aborted flag
- enforce discarding all data for aborted frames on sender and receiver, for interop sanity
- clarified docs - see below

Goals: enable correct use of abort and minimise interop problems with existing
AMQP clients (including proton)

AMPQ requires data for an aborted message to be discarded.  The only case where
it cannot be locally discarded is when part of a message has already been sent
remotely. That is the only case for using abort: if no part of a message has yet
been sent, it is incorrect to use abort.

Note that in a router, an incoming message might be aborted before enough
headers have been acquired to know where to forward it - so silently dropping
some early-aborted messages is unavoidable. Dispatch doesn't promise
frame-by-frame transparency, so discarding a partial message that is aborted
before any of it has been forwarded is valid.

Here's the doc that explains the backwards-compatibility problem:

/**
 * Check if a received delivery has been aborted.
 *
 * An aborted delivery means the sender cannot complete the message and the
 * receiver should discard any data already received. There is nothing further
 * to be done with the delivery: it is pn_delivery_settled() and pn_link_recv()
 * returns ::PN_ABORTED.
 *
 * For backward compatibility with code that does not check pn_delivery_aborted(),
 * the following are true of an aborted delivery:
 *
 * - pn_delivery_partial() is false - old code will not wait forever for more data
 *
 * - pn_delivery_pending() returns 1, not 0 - old code that checks for completion with
 *
 *       if (pn_delivery_pending(d)==0 && !pn_delivery_partial(d))
 *
 *   will not mistake this for successful completion, and will call pn_link_recv()
 *
 * - pn_link_recv() returns ::PN_ABORTED - old code will know there is some kind of error.
 *
 * @see pn_delivery_abort()
 * @param[in] delivery a delivery object
 * @return true if the delivery has been aborted, false otherwise
 */
PN_EXTERN bool pn_delivery_aborted(pn_delivery_t *delivery);