PROTON-1618: c++ tests use test_port.h for listen ports
POSIX: Use bind(0) with SO_REUSEADDR and hold the socket to acquire a port that
can safely be used for listen()

On windows, SO_REUSEADDR has a different meaning, so on windows close the socket
immediately. Theoretically this means another process could steal the port before
the listen() but in practice it seems to work well.

The previous randomize-and-retry approach makes it hard to test the sequence of
events for listen(), since the random retry may cause multiple listen errors
even when it is finally successful.