PROTON-1639: epoll.c proactor error handling bug.
Fixed a bug shown up by the IPV6 test problem. The epoll proactor asserts if a
connect() host:port resolves to a single address, and that address fails to
connect(). In the case of the IPv6 test it was a "network unreachable" error due
to unsupported IPv6, but the problem could occur with any error on connect().