WIP: proton::handle and multithreaded_processor example
proton::handle is a template for an opaque handle to a proton object.

It can be used to get the object's work queue, and the object itself when in a
safe thread context.

The multithreaded_processor example is simple but a realistic use case: a
multi-threaded application wants to process messages in its own threads (not
just proton threads), and accept the message only when processing is complete.

To do that it needs a way to identify deliveries in application threads, and
get the correct work_queue to send the accept() to the correct proton thread.

- implemented proton::handle
- added multithreaded_processor example
- added work_queue to proton::deliery