PROTON-1893: [c] epoll/win_iocp proactor check idle-timeout on open
Set tick_pending in batch_next to make sure there's a tick check after
each CONNECTION_*_OPEN event has been handled.

This generates only 2 extra ticks per connection, and is sufficient:

LOCAL_OPEN is the last chance to change your local time-out before
you've published it to the peer. It is a local event so there's no
danger of hanging if the remote is unresponsive.

REMOTE_OPEN is the only time that the remote time-out can be set.